Move state routing table distance function to trin-state
