+++ trin-core/Cargo.toml
@@ -11,12 +11,10 @@ directories = "3.0"
 env_logger = "0.8.2"
 eth2_ssz = "0.1.2"
 eth2_ssz_derive = "0.1.0"
-ethereum-types = "0.12.0"
 futures = "0.3.13"
 hex = "0.4.3"
 lazy_static = "1.4.0"
 log = "0.4.14"
-num = "0.4.0"
 parking_lot = "0.11.1"
 reqwest = { version = "0.11.0", features = ["blocking"] }
 rlp = "0.5.0"
