+++ trin-state/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 log = "0.4.14"
+num = "0.4.0"
 serde_json = "1.0.59"
 tracing = "0.1.26"
 tracing-subscriber = "0.2.18"
